[major] reformatting functionals
