+++ torchquantum/functional/qubit_unitary.py
@@ -1,85 +1,99 @@
-from typing import Callable, Union, Optional, List, Dict, TYPE_CHECKING
-from ..macro import C_DTYPE, F_DTYPE, ABC, ABC_ARRAY, INV_SQRT2
+import functools
 import torch
 import numpy as np
-from .gate_wrapper import gate_wrapper
+
+from typing import Callable, Union, Optional, List, Dict, TYPE_CHECKING
+from ..macro import C_DTYPE, F_DTYPE, ABC, ABC_ARRAY, INV_SQRT2
+from ..util.utils import pauli_eigs, diag
+from torchpack.utils.logging import logger
+from torchquantum.util import normalize_statevector
+
+from .gate_wrapper import gate_wrapper, apply_unitary_einsum, apply_unitary_bmm
 
 if TYPE_CHECKING:
     from torchquantum.device import QuantumDevice
 else:
     QuantumDevice = None
 
-_hadamard_mat_dict = {
-    "hadamard": torch.tensor(
-        [[INV_SQRT2, INV_SQRT2], [INV_SQRT2, -INV_SQRT2]], dtype=C_DTYPE
-    ),
-    "shadamard": torch.tensor(
-        [
-            [np.cos(np.pi / 8), -np.sin(np.pi / 8)],
-            [np.sin(np.pi / 8), np.cos(np.pi / 8)],
-        ],
-        dtype=C_DTYPE,
-    ),
-    "chadamard": torch.tensor(
-        [
-            [1, 0, 0, 0],
-            [0, 1, 0, 0],
-            [0, 0, INV_SQRT2, INV_SQRT2],
-            [0, 0, INV_SQRT2, -INV_SQRT2],
-        ],
-        dtype=C_DTYPE,
-    ),
+
+def qubitunitary_matrix(params):
+    """Compute unitary matrix for Qubitunitary gate.
+
+    Args:
+        params (torch.Tensor): The unitary matrix.
+
+    Returns:
+        torch.Tensor: The computed unitary matrix.
+
+    """
+    matrix = params.squeeze(0)
+    try:
+        assert matrix.shape[-1] == matrix.shape[-2]
+    except AssertionError as err:
+        logger.exception(f"Operator must be a square matrix.")
+        raise err
+
+    try:
+        U = matrix.cpu().detach().numpy()
+        if matrix.dim() > 2:
+            # batched unitary
+            bsz = matrix.shape[0]
+            assert np.allclose(
+                np.matmul(U, np.transpose(U.conj(), [0, 2, 1])),
+                np.stack([np.identity(U.shape[-1])] * bsz),
+                atol=1e-5,
+            )
+        else:
+            assert np.allclose(
+                np.matmul(U, np.transpose(U.conj(), [1, 0])),
+                np.identity(U.shape[0]),
+                atol=1e-5,
+            )
+    except AssertionError as err:
+        logger.exception(f"Operator must be unitary.")
+        raise err
+
+    return matrix
+
+
+def qubitunitaryfast_matrix(params):
+    """Compute unitary matrix for Qubitunitary fast gate.
+
+    Args:
+        params (torch.Tensor): The unitary matrix.
+
+    Returns:
+        torch.Tensor: The computed unitary matrix.
+
+    """
+    return params.squeeze(0)
+
+
+def qubitunitarystrict_matrix(params):
+    """Compute unitary matrix for Qubitunitary strict gate.
+        Strictly be the unitary.
+
+    Args:
+        params (torch.Tensor): The unitary matrix.
+
+    Returns:
+        torch.Tensor: The computed unitary matrix.
+
+    """
+    params.squeeze(0)
+    mat = params
+    U, Sigma, V = torch.svd(mat)
+    return U.matmul(V)
+
+
+_qubitunitary_mat_dict = {
+    "qubitunitary": qubitunitary_matrix,
+    "qubitunitaryfast": qubitunitaryfast_matrix,
+    "qubitunitarystrict": qubitunitarystrict_matrix,
 }
 
 
-def hadamard(
-    q_device: QuantumDevice,
-    wires: Union[List[int], int],
-    params: torch.Tensor = None,
-    n_wires: int = None,
-    static: bool = False,
-    parent_graph=None,
-    inverse: bool = False,
-    comp_method: str = "bmm",
-):
-    """Perform the hadamard gate.
-
-    Args:
-        q_device (tq.QuantumDevice): The QuantumDevice.
-        wires (Union[List[int], int]): Which qubit(s) to apply the gate.
-        params (torch.Tensor, optional): Parameters (if any) of the gate.
-            Default to None.
-        n_wires (int, optional): Number of qubits the gate is applied to.
-            Default to None.
-        static (bool, optional): Whether use static mode computation.
-            Default to False.
-        parent_graph (tq.QuantumGraph, optional): Parent QuantumGraph of
-            current operation. Default to None.
-        inverse (bool, optional): Whether inverse the gate. Default to False.
-        comp_method (bool, optional): Use 'bmm' or 'einsum' method to perform
-        matrix vector multiplication. Default to 'bmm'.
-
-    Returns:
-        None.
-
-    """
-    name = "hadamard"
-    mat = _hadamard_mat_dict[name]
-    gate_wrapper(
-        name=name,
-        mat=mat,
-        method=comp_method,
-        q_device=q_device,
-        wires=wires,
-        params=params,
-        n_wires=n_wires,
-        static=static,
-        parent_graph=parent_graph,
-        inverse=inverse,
-    )
-
-
-def shadamard(
+def qubitunitary(
     q_device,
     wires,
     params=None,
@@ -89,7 +103,7 @@ def shadamard(
     inverse=False,
     comp_method="bmm",
 ):
-    """Perform the shadamard gate.
+    """Perform the qubitunitary gate.
 
     Args:
         q_device (tq.QuantumDevice): The QuantumDevice.
@@ -110,8 +124,8 @@ def shadamard(
         None.
 
     """
-    name = "shadamard"
-    mat = _hadamard_mat_dict[name]
+    name = "qubitunitary"
+    mat = _qubitunitary_mat_dict[name]
     gate_wrapper(
         name=name,
         mat=mat,
@@ -126,7 +140,7 @@ def shadamard(
     )
 
 
-def chadamard(
+def qubitunitaryfast(
     q_device,
     wires,
     params=None,
@@ -136,7 +150,7 @@ def chadamard(
     inverse=False,
     comp_method="bmm",
 ):
-    """Perform the chadamard gate.
+    """Perform the qubitunitaryfast gate.
 
     Args:
         q_device (tq.QuantumDevice): The QuantumDevice.
@@ -157,10 +171,8 @@ def chadamard(
         None.
 
     """
-
-    name = "chadamard"
-
-    mat = mat_dict[name]
+    name = "qubitunitaryfast"
+    mat = _qubitunitary_mat_dict[name]
     gate_wrapper(
         name=name,
         mat=mat,
@@ -175,6 +187,48 @@ def chadamard(
     )
 
 
-ch = chadamard
-h = hadamard
-sh = shadamard
+def qubitunitarystrict(
+    q_device,
+    wires,
+    params=None,
+    n_wires=None,
+    static=False,
+    parent_graph=None,
+    inverse=False,
+    comp_method="bmm",
+):
+    """Perform the qubitunitarystrict = gate.
+
+    Args:
+        q_device (tq.QuantumDevice): The QuantumDevice.
+        wires (Union[List[int], int]): Which qubit(s) to apply the gate.
+        params (torch.Tensor, optional): Parameters (if any) of the gate.
+            Default to None.
+        n_wires (int, optional): Number of qubits the gate is applied to.
+            Default to None.
+        static (bool, optional): Whether use static mode computation.
+            Default to False.
+        parent_graph (tq.QuantumGraph, optional): Parent QuantumGraph of
+            current operation. Default to None.
+        inverse (bool, optional): Whether inverse the gate. Default to False.
+        comp_method (bool, optional): Use 'bmm' or 'einsum' method to perform
+        matrix vector multiplication. Default to 'bmm'.
+
+    Returns:
+        None.
+
+    """
+    name = "qubitunitarystrict"
+    mat = _qubitunitary_mat_dict[name]
+    gate_wrapper(
+        name=name,
+        mat=mat,
+        method=comp_method,
+        q_device=q_device,
+        wires=wires,
+        params=params,
+        n_wires=n_wires,
+        static=static,
+        parent_graph=parent_graph,
+        inverse=inverse,
+    )
