+++ torchquantum/functional/swap.py
@@ -1,85 +1,54 @@
-from typing import Callable, Union, Optional, List, Dict, TYPE_CHECKING
-from ..macro import C_DTYPE, F_DTYPE, ABC, ABC_ARRAY, INV_SQRT2
+import functools
 import torch
 import numpy as np
-from .gate_wrapper import gate_wrapper
+
+from typing import Callable, Union, Optional, List, Dict, TYPE_CHECKING
+from ..macro import C_DTYPE, F_DTYPE, ABC, ABC_ARRAY, INV_SQRT2
+from ..util.utils import pauli_eigs, diag
+from torchpack.utils.logging import logger
+from torchquantum.util import normalize_statevector
+
+from .gate_wrapper import gate_wrapper, apply_unitary_einsum, apply_unitary_bmm
 
 if TYPE_CHECKING:
     from torchquantum.device import QuantumDevice
 else:
     QuantumDevice = None
 
-_hadamard_mat_dict = {
-    "hadamard": torch.tensor(
-        [[INV_SQRT2, INV_SQRT2], [INV_SQRT2, -INV_SQRT2]], dtype=C_DTYPE
+
+_swap_mat_dict = {
+    "swap": torch.tensor(
+        [[1, 0, 0, 0], [0, 0, 1, 0], [0, 1, 0, 0], [0, 0, 0, 1]], dtype=C_DTYPE
     ),
-    "shadamard": torch.tensor(
-        [
-            [np.cos(np.pi / 8), -np.sin(np.pi / 8)],
-            [np.sin(np.pi / 8), np.cos(np.pi / 8)],
-        ],
-        dtype=C_DTYPE,
-    ),
-    "chadamard": torch.tensor(
+    "sswap": torch.tensor(
         [
             [1, 0, 0, 0],
-            [0, 1, 0, 0],
-            [0, 0, INV_SQRT2, INV_SQRT2],
-            [0, 0, INV_SQRT2, -INV_SQRT2],
+            [0, (1 + 1j) / 2, (1 - 1j) / 2, 0],
+            [0, (1 - 1j) / 2, (1 + 1j) / 2, 0],
+            [0, 0, 0, 1],
         ],
         dtype=C_DTYPE,
+    ),
+    "cswap": torch.tensor(
+        [
+            [1, 0, 0, 0, 0, 0, 0, 0],
+            [0, 1, 0, 0, 0, 0, 0, 0],
+            [0, 0, 1, 0, 0, 0, 0, 0],
+            [0, 0, 0, 1, 0, 0, 0, 0],
+            [0, 0, 0, 0, 1, 0, 0, 0],
+            [0, 0, 0, 0, 0, 0, 1, 0],
+            [0, 0, 0, 0, 0, 1, 0, 0],
+            [0, 0, 0, 0, 0, 0, 0, 1],
+        ],
+        dtype=C_DTYPE,
+    ),
+    "iswap": torch.tensor(
+        [[1, 0, 0, 0], [0, 0, 1j, 0], [0, 1j, 0, 0], [0, 0, 0, 1]], dtype=C_DTYPE
     ),
 }
 
 
-def hadamard(
-    q_device: QuantumDevice,
-    wires: Union[List[int], int],
-    params: torch.Tensor = None,
-    n_wires: int = None,
-    static: bool = False,
-    parent_graph=None,
-    inverse: bool = False,
-    comp_method: str = "bmm",
-):
-    """Perform the hadamard gate.
-
-    Args:
-        q_device (tq.QuantumDevice): The QuantumDevice.
-        wires (Union[List[int], int]): Which qubit(s) to apply the gate.
-        params (torch.Tensor, optional): Parameters (if any) of the gate.
-            Default to None.
-        n_wires (int, optional): Number of qubits the gate is applied to.
-            Default to None.
-        static (bool, optional): Whether use static mode computation.
-            Default to False.
-        parent_graph (tq.QuantumGraph, optional): Parent QuantumGraph of
-            current operation. Default to None.
-        inverse (bool, optional): Whether inverse the gate. Default to False.
-        comp_method (bool, optional): Use 'bmm' or 'einsum' method to perform
-        matrix vector multiplication. Default to 'bmm'.
-
-    Returns:
-        None.
-
-    """
-    name = "hadamard"
-    mat = _hadamard_mat_dict[name]
-    gate_wrapper(
-        name=name,
-        mat=mat,
-        method=comp_method,
-        q_device=q_device,
-        wires=wires,
-        params=params,
-        n_wires=n_wires,
-        static=static,
-        parent_graph=parent_graph,
-        inverse=inverse,
-    )
-
-
-def shadamard(
+def swap(
     q_device,
     wires,
     params=None,
@@ -89,7 +58,7 @@ def shadamard(
     inverse=False,
     comp_method="bmm",
 ):
-    """Perform the shadamard gate.
+    """Perform the swap gate.
 
     Args:
         q_device (tq.QuantumDevice): The QuantumDevice.
@@ -110,8 +79,8 @@ def shadamard(
         None.
 
     """
-    name = "shadamard"
-    mat = _hadamard_mat_dict[name]
+    name = "swap"
+    mat = _swap_mat_dict[name]
     gate_wrapper(
         name=name,
         mat=mat,
@@ -126,7 +95,7 @@ def shadamard(
     )
 
 
-def chadamard(
+def sswap(
     q_device,
     wires,
     params=None,
@@ -136,7 +105,7 @@ def chadamard(
     inverse=False,
     comp_method="bmm",
 ):
-    """Perform the chadamard gate.
+    """Perform the sswap gate.
 
     Args:
         q_device (tq.QuantumDevice): The QuantumDevice.
@@ -157,10 +126,8 @@ def chadamard(
         None.
 
     """
-
-    name = "chadamard"
-
-    mat = mat_dict[name]
+    name = "sswap"
+    mat = _swap_mat_dict[name]
     gate_wrapper(
         name=name,
         mat=mat,
@@ -175,6 +142,95 @@ def chadamard(
     )
 
 
-ch = chadamard
-h = hadamard
-sh = shadamard
+def cswap(
+    q_device,
+    wires,
+    params=None,
+    n_wires=None,
+    static=False,
+    parent_graph=None,
+    inverse=False,
+    comp_method="bmm",
+):
+    """Perform the cswap gate.
+
+    Args:
+        q_device (tq.QuantumDevice): The QuantumDevice.
+        wires (Union[List[int], int]): Which qubit(s) to apply the gate.
+        params (torch.Tensor, optional): Parameters (if any) of the gate.
+            Default to None.
+        n_wires (int, optional): Number of qubits the gate is applied to.
+            Default to None.
+        static (bool, optional): Whether use static mode computation.
+            Default to False.
+        parent_graph (tq.QuantumGraph, optional): Parent QuantumGraph of
+            current operation. Default to None.
+        inverse (bool, optional): Whether inverse the gate. Default to False.
+        comp_method (bool, optional): Use 'bmm' or 'einsum' method to perform
+        matrix vector multiplication. Default to 'bmm'.
+
+    Returns:
+        None.
+
+    """
+    name = "cswap"
+    mat = _swap_mat_dict[name]
+    gate_wrapper(
+        name=name,
+        mat=mat,
+        method=comp_method,
+        q_device=q_device,
+        wires=wires,
+        params=params,
+        n_wires=n_wires,
+        static=static,
+        parent_graph=parent_graph,
+        inverse=inverse,
+    )
+
+
+def iswap(
+    q_device,
+    wires,
+    params=None,
+    n_wires=None,
+    static=False,
+    parent_graph=None,
+    inverse=False,
+    comp_method="bmm",
+):
+    """Perform the iswap gate.
+
+    Args:
+        q_device (tq.QuantumDevice): The QuantumDevice.
+        wires (Union[List[int], int]): Which qubit(s) to apply the gate.
+        params (torch.Tensor, optional): Parameters (if any) of the gate.
+            Default to None.
+        n_wires (int, optional): Number of qubits the gate is applied to.
+            Default to None.
+        static (bool, optional): Whether use static mode computation.
+            Default to False.
+        parent_graph (tq.QuantumGraph, optional): Parent QuantumGraph of
+            current operation. Default to None.
+        inverse (bool, optional): Whether inverse the gate. Default to False.
+        comp_method (bool, optional): Use 'bmm' or 'einsum' method to perform
+        matrix vector multiplication. Default to 'bmm'.
+
+    Returns:
+        None.
+
+    """
+    name = "iswap"
+    mat = _swap_mat_dict[name]
+    gate_wrapper(
+        name=name,
+        mat=mat,
+        method=comp_method,
+        q_device=q_device,
+        wires=wires,
+        params=params,
+        n_wires=n_wires,
+        static=static,
+        parent_graph=parent_graph,
+        inverse=inverse,
+    )
