+++ torchquantum/functional/u1.py
@@ -1,85 +1,84 @@
-from typing import Callable, Union, Optional, List, Dict, TYPE_CHECKING
-from ..macro import C_DTYPE, F_DTYPE, ABC, ABC_ARRAY, INV_SQRT2
+import functools
 import torch
 import numpy as np
-from .gate_wrapper import gate_wrapper
+
+from typing import Callable, Union, Optional, List, Dict, TYPE_CHECKING
+from ..macro import C_DTYPE, F_DTYPE, ABC, ABC_ARRAY, INV_SQRT2
+from ..util.utils import pauli_eigs, diag
+from torchpack.utils.logging import logger
+from torchquantum.util import normalize_statevector
+
+from .gate_wrapper import gate_wrapper, apply_unitary_einsum, apply_unitary_bmm
 
 if TYPE_CHECKING:
     from torchquantum.device import QuantumDevice
 else:
     QuantumDevice = None
 
-_hadamard_mat_dict = {
-    "hadamard": torch.tensor(
-        [[INV_SQRT2, INV_SQRT2], [INV_SQRT2, -INV_SQRT2]], dtype=C_DTYPE
-    ),
-    "shadamard": torch.tensor(
+
+def u1_matrix(params):
+    """Compute unitary matrix for U1 gate.
+
+    Args:
+        params (torch.Tensor): The rotation angle.
+
+    Returns:
+        torch.Tensor: The computed unitary matrix.
+
+    """
+    phi = params.type(C_DTYPE)
+    exp = torch.exp(1j * phi)
+
+    return torch.stack(
         [
-            [np.cos(np.pi / 8), -np.sin(np.pi / 8)],
-            [np.sin(np.pi / 8), np.cos(np.pi / 8)],
+            torch.cat(
+                [
+                    torch.ones(exp.shape, device=params.device),
+                    torch.zeros(exp.shape, device=params.device),
+                ],
+                dim=-1,
+            ),
+            torch.cat([torch.zeros(exp.shape, device=params.device), exp], dim=-1),
         ],
-        dtype=C_DTYPE,
-    ),
-    "chadamard": torch.tensor(
-        [
-            [1, 0, 0, 0],
-            [0, 1, 0, 0],
-            [0, 0, INV_SQRT2, INV_SQRT2],
-            [0, 0, INV_SQRT2, -INV_SQRT2],
-        ],
-        dtype=C_DTYPE,
-    ),
+        dim=-2,
+    ).squeeze(0)
+
+
+def cu1_matrix(params):
+    """Compute unitary matrix for CU1 gate.
+
+    Args:
+        params (torch.Tensor): The rotation angle.
+
+    Returns:
+        torch.Tensor: The computed unitary matrix.
+
+    """
+    phi = params.type(C_DTYPE)
+    exp = torch.exp(1j * phi)
+
+    matrix = (
+        torch.tensor(
+            [[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 1, 0], [0, 0, 0, 0]],
+            dtype=C_DTYPE,
+            device=params.device,
+        )
+        .unsqueeze(0)
+        .repeat(phi.shape[0], 1, 1)
+    )
+
+    matrix[:, 3, 3] = exp
+
+    return matrix.squeeze(0)
+
+
+_u1_mat_dict = {
+    "u1": u1_matrix,
+    "cu1": cu1_matrix,
 }
 
 
-def hadamard(
-    q_device: QuantumDevice,
-    wires: Union[List[int], int],
-    params: torch.Tensor = None,
-    n_wires: int = None,
-    static: bool = False,
-    parent_graph=None,
-    inverse: bool = False,
-    comp_method: str = "bmm",
-):
-    """Perform the hadamard gate.
-
-    Args:
-        q_device (tq.QuantumDevice): The QuantumDevice.
-        wires (Union[List[int], int]): Which qubit(s) to apply the gate.
-        params (torch.Tensor, optional): Parameters (if any) of the gate.
-            Default to None.
-        n_wires (int, optional): Number of qubits the gate is applied to.
-            Default to None.
-        static (bool, optional): Whether use static mode computation.
-            Default to False.
-        parent_graph (tq.QuantumGraph, optional): Parent QuantumGraph of
-            current operation. Default to None.
-        inverse (bool, optional): Whether inverse the gate. Default to False.
-        comp_method (bool, optional): Use 'bmm' or 'einsum' method to perform
-        matrix vector multiplication. Default to 'bmm'.
-
-    Returns:
-        None.
-
-    """
-    name = "hadamard"
-    mat = _hadamard_mat_dict[name]
-    gate_wrapper(
-        name=name,
-        mat=mat,
-        method=comp_method,
-        q_device=q_device,
-        wires=wires,
-        params=params,
-        n_wires=n_wires,
-        static=static,
-        parent_graph=parent_graph,
-        inverse=inverse,
-    )
-
-
-def shadamard(
+def u1(
     q_device,
     wires,
     params=None,
@@ -89,7 +88,7 @@ def shadamard(
     inverse=False,
     comp_method="bmm",
 ):
-    """Perform the shadamard gate.
+    """Perform the u1 gate.
 
     Args:
         q_device (tq.QuantumDevice): The QuantumDevice.
@@ -110,56 +109,7 @@ def shadamard(
         None.
 
     """
-    name = "shadamard"
-    mat = _hadamard_mat_dict[name]
-    gate_wrapper(
-        name=name,
-        mat=mat,
-        method=comp_method,
-        q_device=q_device,
-        wires=wires,
-        params=params,
-        n_wires=n_wires,
-        static=static,
-        parent_graph=parent_graph,
-        inverse=inverse,
-    )
-
-
-def chadamard(
-    q_device,
-    wires,
-    params=None,
-    n_wires=None,
-    static=False,
-    parent_graph=None,
-    inverse=False,
-    comp_method="bmm",
-):
-    """Perform the chadamard gate.
-
-    Args:
-        q_device (tq.QuantumDevice): The QuantumDevice.
-        wires (Union[List[int], int]): Which qubit(s) to apply the gate.
-        params (torch.Tensor, optional): Parameters (if any) of the gate.
-            Default to None.
-        n_wires (int, optional): Number of qubits the gate is applied to.
-            Default to None.
-        static (bool, optional): Whether use static mode computation.
-            Default to False.
-        parent_graph (tq.QuantumGraph, optional): Parent QuantumGraph of
-            current operation. Default to None.
-        inverse (bool, optional): Whether inverse the gate. Default to False.
-        comp_method (bool, optional): Use 'bmm' or 'einsum' method to perform
-        matrix vector multiplication. Default to 'bmm'.
-
-    Returns:
-        None.
-
-    """
-
-    name = "chadamard"
-
+    name = "u1"
     mat = mat_dict[name]
     gate_wrapper(
         name=name,
@@ -175,6 +125,53 @@ def chadamard(
     )
 
 
-ch = chadamard
-h = hadamard
-sh = shadamard
+def cu1(
+    q_device,
+    wires,
+    params=None,
+    n_wires=None,
+    static=False,
+    parent_graph=None,
+    inverse=False,
+    comp_method="bmm",
+):
+    """Perform the cu1 gate.
+
+    Args:
+        q_device (tq.QuantumDevice): The QuantumDevice.
+        wires (Union[List[int], int]): Which qubit(s) to apply the gate.
+        params (torch.Tensor, optional): Parameters (if any) of the gate.
+            Default to None.
+        n_wires (int, optional): Number of qubits the gate is applied to.
+            Default to None.
+        static (bool, optional): Whether use static mode computation.
+            Default to False.
+        parent_graph (tq.QuantumGraph, optional): Parent QuantumGraph of
+            current operation. Default to None.
+        inverse (bool, optional): Whether inverse the gate. Default to False.
+        comp_method (bool, optional): Use 'bmm' or 'einsum' method to perform
+        matrix vector multiplication. Default to 'bmm'.
+
+    Returns:
+        None.
+
+    """
+    name = "cu1"
+    mat = mat_dict[name]
+    gate_wrapper(
+        name=name,
+        mat=mat,
+        method=comp_method,
+        q_device=q_device,
+        wires=wires,
+        params=params,
+        n_wires=n_wires,
+        static=static,
+        parent_graph=parent_graph,
+        inverse=inverse,
+    )
+
+
+cp = cu1
+cr = cu1
+cphase = cu1
