+++ torchquantum/functional/u3.py
@@ -1,85 +1,125 @@
-from typing import Callable, Union, Optional, List, Dict, TYPE_CHECKING
-from ..macro import C_DTYPE, F_DTYPE, ABC, ABC_ARRAY, INV_SQRT2
+import functools
 import torch
 import numpy as np
-from .gate_wrapper import gate_wrapper
+
+from typing import Callable, Union, Optional, List, Dict, TYPE_CHECKING
+from ..macro import C_DTYPE, F_DTYPE, ABC, ABC_ARRAY, INV_SQRT2
+from ..util.utils import pauli_eigs, diag
+from torchpack.utils.logging import logger
+from torchquantum.util import normalize_statevector
+
+from .gate_wrapper import gate_wrapper, apply_unitary_einsum, apply_unitary_bmm
 
 if TYPE_CHECKING:
     from torchquantum.device import QuantumDevice
 else:
     QuantumDevice = None
 
-_hadamard_mat_dict = {
-    "hadamard": torch.tensor(
-        [[INV_SQRT2, INV_SQRT2], [INV_SQRT2, -INV_SQRT2]], dtype=C_DTYPE
-    ),
-    "shadamard": torch.tensor(
+
+def cu_matrix(params):
+    """Compute unitary matrix for CU gate.
+    Args:
+        params (torch.Tensor): The rotation angle.
+    Returns:
+        torch.Tensor: The computed unitary matrix.
+    """
+    theta = params[:, 0].unsqueeze(dim=-1).type(C_DTYPE)
+    phi = params[:, 1].unsqueeze(dim=-1).type(C_DTYPE)
+    lam = params[:, 2].unsqueeze(dim=-1).type(C_DTYPE)
+    gamma = params[:, 3].unsqueeze(dim=-1).type(C_DTYPE)
+
+    co = torch.cos(theta / 2)
+    si = torch.sin(theta / 2)
+
+    matrix = (
+        torch.tensor(
+            [[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, 0], [0, 0, 0, 0]],
+            dtype=C_DTYPE,
+            device=params.device,
+        )
+        .unsqueeze(0)
+        .repeat(phi.shape[0], 1, 1)
+    )
+
+    matrix[:, 2, 2] = co * torch.exp(1j * gamma)
+    matrix[:, 2, 3] = -si * torch.exp(1j * (lam + gamma))
+    matrix[:, 3, 2] = si * torch.exp(1j * (phi + gamma))
+    matrix[:, 3, 3] = co * torch.exp(1j * (phi + lam + gamma))
+
+    return matrix.squeeze(0)
+
+
+def u3_matrix(params):
+    """Compute unitary matrix for U3 gate.
+
+    Args:
+        params (torch.Tensor): The rotation angle.
+
+    Returns:
+        torch.Tensor: The computed unitary matrix.
+
+    """
+    theta = params[:, 0].unsqueeze(dim=-1).type(C_DTYPE)
+    phi = params[:, 1].unsqueeze(dim=-1).type(C_DTYPE)
+    lam = params[:, 2].unsqueeze(dim=-1).type(C_DTYPE)
+
+    co = torch.cos(theta / 2)
+    si = torch.sin(theta / 2)
+
+    return torch.stack(
         [
-            [np.cos(np.pi / 8), -np.sin(np.pi / 8)],
-            [np.sin(np.pi / 8), np.cos(np.pi / 8)],
+            torch.cat([co, -si * torch.exp(1j * lam)], dim=-1),
+            torch.cat(
+                [si * torch.exp(1j * phi), co * torch.exp(1j * (phi + lam))], dim=-1
+            ),
         ],
-        dtype=C_DTYPE,
-    ),
-    "chadamard": torch.tensor(
-        [
-            [1, 0, 0, 0],
-            [0, 1, 0, 0],
-            [0, 0, INV_SQRT2, INV_SQRT2],
-            [0, 0, INV_SQRT2, -INV_SQRT2],
-        ],
-        dtype=C_DTYPE,
-    ),
+        dim=-2,
+    ).squeeze(0)
+
+
+def cu3_matrix(params):
+    """Compute unitary matrix for CU3 gate.
+
+    Args:
+        params (torch.Tensor): The rotation angle.
+
+    Returns:
+        torch.Tensor: The computed unitary matrix.
+
+    """
+    theta = params[:, 0].unsqueeze(dim=-1).type(C_DTYPE)
+    phi = params[:, 1].unsqueeze(dim=-1).type(C_DTYPE)
+    lam = params[:, 2].unsqueeze(dim=-1).type(C_DTYPE)
+
+    co = torch.cos(theta / 2)
+    si = torch.sin(theta / 2)
+
+    matrix = (
+        torch.tensor(
+            [[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, 0], [0, 0, 0, 0]],
+            dtype=C_DTYPE,
+            device=params.device,
+        )
+        .unsqueeze(0)
+        .repeat(phi.shape[0], 1, 1)
+    )
+
+    matrix[:, 2, 2] = co
+    matrix[:, 2, 3] = -si * torch.exp(1j * lam)
+    matrix[:, 3, 2] = si * torch.exp(1j * phi)
+    matrix[:, 3, 3] = co * torch.exp(1j * (phi + lam))
+
+    return matrix.squeeze(0)
+
+
+_u3_mat_dict = {
+    "u3": u3_matrix,
+    "cu3": cu3_matrix,
+    "cu": cu_matrix,
 }
 
 
-def hadamard(
-    q_device: QuantumDevice,
-    wires: Union[List[int], int],
-    params: torch.Tensor = None,
-    n_wires: int = None,
-    static: bool = False,
-    parent_graph=None,
-    inverse: bool = False,
-    comp_method: str = "bmm",
-):
-    """Perform the hadamard gate.
-
-    Args:
-        q_device (tq.QuantumDevice): The QuantumDevice.
-        wires (Union[List[int], int]): Which qubit(s) to apply the gate.
-        params (torch.Tensor, optional): Parameters (if any) of the gate.
-            Default to None.
-        n_wires (int, optional): Number of qubits the gate is applied to.
-            Default to None.
-        static (bool, optional): Whether use static mode computation.
-            Default to False.
-        parent_graph (tq.QuantumGraph, optional): Parent QuantumGraph of
-            current operation. Default to None.
-        inverse (bool, optional): Whether inverse the gate. Default to False.
-        comp_method (bool, optional): Use 'bmm' or 'einsum' method to perform
-        matrix vector multiplication. Default to 'bmm'.
-
-    Returns:
-        None.
-
-    """
-    name = "hadamard"
-    mat = _hadamard_mat_dict[name]
-    gate_wrapper(
-        name=name,
-        mat=mat,
-        method=comp_method,
-        q_device=q_device,
-        wires=wires,
-        params=params,
-        n_wires=n_wires,
-        static=static,
-        parent_graph=parent_graph,
-        inverse=inverse,
-    )
-
-
-def shadamard(
+def u3(
     q_device,
     wires,
     params=None,
@@ -89,7 +129,7 @@ def shadamard(
     inverse=False,
     comp_method="bmm",
 ):
-    """Perform the shadamard gate.
+    """Perform the u3 gate.
 
     Args:
         q_device (tq.QuantumDevice): The QuantumDevice.
@@ -110,8 +150,8 @@ def shadamard(
         None.
 
     """
-    name = "shadamard"
-    mat = _hadamard_mat_dict[name]
+    name = "u3"
+    mat = _u3_mat_dict[name]
     gate_wrapper(
         name=name,
         mat=mat,
@@ -126,7 +166,7 @@ def shadamard(
     )
 
 
-def chadamard(
+def cu(
     q_device,
     wires,
     params=None,
@@ -136,7 +176,51 @@ def chadamard(
     inverse=False,
     comp_method="bmm",
 ):
-    """Perform the chadamard gate.
+    """Perform the cu gate.
+    Args:
+        q_device (tq.QuantumDevice): The QuantumDevice.
+        wires (Union[List[int], int]): Which qubit(s) to apply the gate.
+        params (torch.Tensor, optional): Parameters (if any) of the gate.
+            Default to None.
+        n_wires (int, optional): Number of qubits the gate is applied to.
+            Default to None.
+        static (bool, optional): Whether use static mode computation.
+            Default to False.
+        parent_graph (tq.QuantumGraph, optional): Parent QuantumGraph of
+            current operation. Default to None.
+        inverse (bool, optional): Whether inverse the gate. Default to False.
+        comp_method (bool, optional): Use 'bmm' or 'einsum' method to perform
+        matrix vector multiplication. Default to 'bmm'.
+    Returns:
+        None.
+    """
+    name = "cu"
+    mat = _u3_mat_dict[name]
+    gate_wrapper(
+        name=name,
+        mat=mat,
+        method=comp_method,
+        q_device=q_device,
+        wires=wires,
+        params=params,
+        n_wires=n_wires,
+        static=static,
+        parent_graph=parent_graph,
+        inverse=inverse,
+    )
+
+
+def cu3(
+    q_device,
+    wires,
+    params=None,
+    n_wires=None,
+    static=False,
+    parent_graph=None,
+    inverse=False,
+    comp_method="bmm",
+):
+    """Perform the cu3 gate.
 
     Args:
         q_device (tq.QuantumDevice): The QuantumDevice.
@@ -157,10 +241,8 @@ def chadamard(
         None.
 
     """
-
-    name = "chadamard"
-
-    mat = mat_dict[name]
+    name = "cu3"
+    mat = _u3_mat_dict[name]
     gate_wrapper(
         name=name,
         mat=mat,
@@ -175,6 +257,4 @@ def chadamard(
     )
 
 
-ch = chadamard
-h = hadamard
-sh = shadamard
+u = u3
