+++ torchquantum/operator/standard_gates/phase_shift.py
@@ -4,7 +4,7 @@ from torchquantum.macro import C_DTYPE
 import torchquantum as tq
 import torch
 from torchquantum.functional import mat_dict
-import torchquantum.functional.functionals as tqf
+import torchquantum.functional as tqf
 
 
 class PhaseShift(DiagonalOperation, metaclass=ABCMeta):
